+++ site/src/app/routes/_page/docs/collaboration/history.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { CollaborationHistoryRoute } from "../../../../../routes/docs/DocsRoute";
-
-export const Route = createFileRoute("/_page/docs/collaboration/history")({
-  component: CollaborationHistoryRoute,
-});